% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuralfable.R
\name{neuralfable}
\alias{neuralfable}
\title{Neuralfable models}
\usage{
neuralfable(formula, ...)
}
\arguments{
\item{formula}{A neuralfable model formula. This encompasses all models exposed by neuralfables (see details).}

\item{...}{Additional arguments (see details).}
}
\value{
A specified model, analogous to other model objects within fable/fabletools.
}
\description{
Neurafable model class
}
\details{
Use the method special to specify a method - available methods are **"elm", "mlp", "garbage", "esn"**,
with default being **"mlp"**.
}
